requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.2
gunicorn==20.1.0
h11==0.14.0
hurry.filesize==0.9
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
pathlib==1.0.1
pydantic==1.10.2
python-multipart==0.0.5
pytube==12.1.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
